v0.3.1143 follow-on — SCALE-SEAM ⓺ ⓻ ⓼, preflight / types / groups
Eleven methods out of client.ts (1,584 → 1,512). No extra withX().
TypeRow / TypeDetail / GroupRow / AssemblyRow moved to types.ts.
