+++ apps/web/src/api/authoring.ts
@@ -6,6 +6,12 @@
  *  authoring chain?* Property-override layers sat immediately below in `client.ts` and did
  *  **not** come — those compose properties, they do not write recipes.
  *
+ *  SCALE-SEAM ⓻ adds family types — *what families can I place?* List, inspector,
+ *  create, edit, material-layer set. Groups sat below and did **not** come with that slice.
+ *
+ *  SCALE-SEAM ⓼ adds groups and assemblies — *how are these elements grouped?*
+ *  List, inspector, create group/assembly, parametric array. Detailing stayed.
+ *
  *  First extraction of roadmap SCALE-SEAM. `client.ts` was measured at 4,956 lines with 152 commits
  *  in a fortnight and 631 methods on one class: it had to be opened to add any endpoint, so every
  *  change to it competed with every other change. The server solved this long ago by splitting into
@@ -21,7 +27,7 @@
  *  sibling's private member — that has to move into HttpCore first.
  */
 import { HttpCore } from "./httpCore";
-import type { EditMacro } from "./types";
+import type { AssemblyRow, EditMacro, GroupRow, TypeDetail, TypeRow } from "./types";
 
 type Ctor<T> = new (...args: any[]) => T;
 
@@ -259,6 +265,55 @@ export function withAuthoring<TBase extends Ctor<HttpCore>>(Base: TBase) {
       return this.json<Record<string, unknown>>(
         `/projects/${pid}/macros/${encodeURIComponent(macroId)}/run`,
         { method: "POST", body: JSON.stringify({ args, publish: opts?.publish ?? false, base_source: opts?.baseSource ?? null }) });
+    }
+    /** Placeable types ("families") in the project's source IFC, for the place-family picker and the
+     *  type browser. Carries PredefinedType + how many occurrences reference each type. */
+    types(pid: string) {
+      return this.json<{ types: TypeRow[] }>(`/projects/${pid}/types`);
+    }
+    /** W10-1 type inspector: class, predefined, box dims, type Psets, material layers, occurrences. */
+    typeDetail(pid: string, typeGuid: string) {
+      return this.json<TypeDetail>(`/projects/${pid}/types/${encodeURIComponent(typeGuid)}`);
+    }
+    /** W10-1: author a custom family type (class + optional [w,d,h] box + PredefinedType + type Psets).
+     *  Returns the new type GUID in `changed`. Versioned + GUID-stable via the /edit recipe path. */
+    createType(pid: string, ifc_class: string, name: string, dims?: [number, number, number] | null,
+               predefined?: string | null, psets?: Record<string, Record<string, unknown>> | null,
+               publish = true) {
+      return this.editIfc(pid, "create_type", { ifc_class, name, dims, predefined, psets }, publish);
+    }
+    /** W10-1: edit a type's params. Changing `dims` propagates to EVERY placed occurrence at once
+     *  (shared RepresentationMap), GUID-stable — no re-placement. */
+    editType(pid: string, type_guid: string, patch: { name?: string; dims?: [number, number, number];
+               predefined?: string; psets?: Record<string, Record<string, unknown>> }, publish = true) {
+      return this.editIfc(pid, "edit_type_params", { type_guid, ...patch }, publish);
+    }
+    /** W10-1: give a type an ordered IfcMaterialLayerSet ([{material, thickness(m)}]); occurrences inherit. */
+    assignMaterialSet(pid: string, type_guid: string,
+                      layers: { material: string; thickness: number }[], publish = true) {
+      return this.editIfc(pid, "assign_material_set", { type_guid, layers }, publish);
+    }
+    /** W10-3: every IfcGroup (named set) and IfcElementAssembly (part-of whole) with member counts. */
+    groups(pid: string) {
+      return this.json<{ groups: GroupRow[]; assemblies: AssemblyRow[] }>(`/projects/${pid}/groups`);
+    }
+    /** W10-3 inspector: the members/parts of one group or assembly. */
+    groupDetail(pid: string, guid: string) {
+      return this.json<{ guid: string; kind: "group" | "assembly"; name: string; member_count: number;
+        members: { guid: string; name: string; ifc_class: string }[] }>(
+        `/projects/${pid}/groups/${encodeURIComponent(guid)}`);
+    }
+    /** W10-3: author an IfcGroup (named set) over the given element GUIDs (re-using a name adds to it). */
+    createGroup(pid: string, name: string, guids: string[], publish = true) {
+      return this.editIfc(pid, "create_group", { name, guids }, publish);
+    }
+    /** W10-3: aggregate the given elements into an IfcElementAssembly (a real part-of whole). */
+    createAssembly(pid: string, name: string, guids: string[], predefined?: string | null, publish = true) {
+      return this.editIfc(pid, "create_assembly", { name, guids, predefined }, publish);
+    }
+    /** W10-3: rectangular parametric array — nx×ny copies at pitch (dx,dy) m (dz per column). */
+    arrayElement(pid: string, guid: string, nx: number, ny: number, dx: number, dy: number, dz = 0, publish = true) {
+      return this.editIfc(pid, "array_element", { guid, nx, ny, dx, dy, dz }, publish);
     }
   };
 }
