+++ apps/web/src/api/surface.test.ts
@@ -210,6 +210,9 @@ describe("the API client's public surface", () => {
       "integrations", "license", "capabilities",                    // 23 deploy entitle
       "siteContext", "parcelAnalyze", "parcelsScreen",              // 24 land around site
       "pins",                                                       // 25 field marks
+      "preflight",                                                   // 26 issuance gate
+      "types", "typeDetail", "createType", "editType",               // 27 families
+      "groups", "createGroup", "createAssembly", "arrayElement",     // 28 groups
       // ⑫ moved — one contiguous /connections run; a silent loss here is the admin connections screen
       "connections", "createConnection", "testConnection", "connectionTables",
       "syncProcore", "pushProcore", "syncSchedules",
