+++ apps/web/src/shell/roadmapLanes.test.ts
@@ -79,7 +79,8 @@ const OPEN = LINES.slice(0, GATED_AT === -1 ? LINES.length : GATED_AT);
 // Widened to ⓴ on 2026-09-01, in the follow-on that SHIPS ⓱ ⓲ ⓳ without a version bump.
 // Widened to ⓷ on 2026-09-01, in the follow-on that SHIPS ⓴ ⓵ ⓶ without a version bump.
 // Widened to ⓺ on 2026-09-01, in the follow-on that SHIPS ⓷ ⓸ ⓹ without a version bump.
-const MARKS = "①②③④⑤⑥⑦⑧⑨⑩⑪⑫⑬⑭⑮⑯⑰⑱⑲⑳㉑㉒㉓㉔㉕㉖㉗㉘㉙㉚㉛㉜㉝㉞㉟㊱㊲㊳㊴㊵㊶㊷㊸㊹㊺㊻㊼㊽㊾㊿❶❷❸❹❺❻❼❽❾❿⓫⓬⓭⓮⓯⓰⓱⓲⓳⓴⓵⓶⓷⓸⓹⓺";
+// Widened to ⓽ on 2026-09-02, in the follow-on that SHIPS ⓺ ⓻ ⓼ without a version bump.
+const MARKS = "①②③④⑤⑥⑦⑧⑨⑩⑪⑫⑬⑭⑮⑯⑰⑱⑲⑳㉑㉒㉓㉔㉕㉖㉗㉘㉙㉚㉛㉜㉝㉞㉟㊱㊲㊳㊴㊵㊶㊷㊸㊹㊺㊻㊼㊽㊾㊿❶❷❸❹❺❻❼❽❾❿⓫⓬⓭⓮⓯⓰⓱⓲⓳⓴⓵⓶⓷⓸⓹⓺⓻⓼⓽";
 
 /**
  * One source for the marker vocabulary, because there were **two** and they had already drifted.
