+++ docs/roadmap.md
@@ -1188,7 +1188,7 @@ two rows share a path, so two agents in different rows cannot collide.
 | **F · Docs & demo** | `README.md`, `docs/`, `apps/web/src/demo/` | keep the shipped surface honest (below) — no coded items. **`demoData.test.ts` now gates the shell's startup endpoints**; re-run `build_demo_data.py` and that test after adding one |
 | **G · API surface** | `services/api/src/aec_api/routers/`, `main.py` | no standalone items: **every lane routes its own work**, which is why this is a lane rather than a shared file |
 | **H · Registers** | `services/api/modules/*/module.json` | — |
-| **I · API client** | `apps/web/src/api/` | SCALE-SEAM ⓺ *(the only open slice; ②–⓹ have shipped. This cell named ⑬–⑳ until 2026-08-24 — eight slices whose extractions had already landed — because the item regex could not see `㉒` at all, so nothing required this row to be right)* |
+| **I · API client** | `apps/web/src/api/` | SCALE-SEAM ⓽ *(the only open slice; ②–⓼ have shipped. This cell named ⑬–⑳ until 2026-08-24 — eight slices whose extractions had already landed — because the item regex could not see `㉒` at all, so nothing required this row to be right)* |
 | **J · Build & tooling** | `apps/web/scripts/`, `apps/web/vite.config.ts`, `apps/web/src/style.css`, `apps/web/src/tooling/`, `services/api/test_file_sizes.py`, `services/api/run_tests.py` | R39-TSC-CACHE *(local typecheck once diverged from CI; cause unknown, prior explanation retracted — an OBSERVATION, not a defect with a known fix. Read the entry before "fixing" it: the proposed fix is named there and rejected)* |
 
 **Parked — not available to pick up.** These are decisions or multi-release commitments, listed so
@@ -3061,7 +3061,21 @@ verbs, with a command bar as the escape hatch to everything); and **role-shaped 
 
 ## 🧱 Decomposition & reliability carry-overs (interleave one per few releases)
 
-- ◧ ⭐ **SCALE-SEAM ⓺ — `client.ts` is no longer a god-file, but the split is not finished.** *(②–⓹ have shipped, **⓷–⓹ in the v0.3.1143 follow-on** — no version bump, tag lag already at the `test_release_current` bound)*
+- ◧ ⭐ **SCALE-SEAM ⓽ — `client.ts` is no longer a god-file, but the split is not finished.** *(②–⓼ have shipped, **⓺–⓼ in the v0.3.1143 follow-on** — no version bump, tag lag already at the `test_release_current` bound)*
+  **⓼ took groups and assemblies out** (five methods) into the existing
+  `apps/web/src/api/authoring.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by
+  what they ANSWER: *how are these elements grouped?* **Detailing did NOT come.**
+
+  **⓻ took family types out** (five methods) into the existing
+  `apps/web/src/api/authoring.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by
+  what they ANSWER: *what families can I place?* **Groups did NOT come with ⓻.**
+
+  **⓺ took the issuance gate out** (one method) into the existing
+  `apps/web/src/api/documents.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by
+  what they ANSWER: *can this package go out?* **Hero upload did NOT come.**
+  **`MARKS` was widened to ⓽ on the way out.** **The per-file pin moved 1,584 → 1,512**
+  across ⓺+⓻+⓼.
+
   **⓹ took field pins out** (one method) into the existing
   `apps/web/src/api/topics.ts` mixin — **no extra wrapper on `ApiClient`**. Grouped by what
   they ANSWER: *where are the field marks?* **⑳ left pins behind (not `/topics`).**
